+++ src/analyzer.ts
@@ -42,6 +42,7 @@ function processConfigXML(folder: string) {
         config.preferences[pref['@_name']] = pref['@_value'];
       }
     }
+    if (!widget.platform) return config;
     for (const platform of widget.platform) {
       if (platform['@_name'] == 'android' && platform.preference) {
         for (const pref of platform.preference) {
@@ -178,6 +179,7 @@ export function remotePackages(): Array<string> {
 
 export function deprecatedPackages(packages: any): Array<any> {
   const result = [];
+  if (!packages) return result;
   for (const library of Object.keys(packages)) {
     if (packages[library].deprecated) {
       result.push({ name: library, message: packages[library].deprecated });